refactor : 생성자 주입으로 변경 및 테스트 안돌아가는 문제 수정
