+++ src/test/java/roomescape/controller/AdminApiTest.java
@@ -1,4 +1,4 @@
-package roomescape.admin;
+package roomescape.controller;
 
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.BeforeEach;
@@ -10,7 +10,7 @@ import org.springframework.test.annotation.DirtiesContext;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_CLASS)
-public class ApiTest {
+public class AdminApiTest {
 
     @LocalServerPort
     int port;
