+++ src/test/java/roomescape/controller/ReservationsApiTest.java
@@ -15,7 +15,7 @@ import static org.hamcrest.Matchers.is;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_EACH_TEST_METHOD)
-public class ApiTest {
+public class ReservationsApiTest {
 
     @LocalServerPort
     int port;
