+++ src/test/java/roomescape/controller/TimesApiTest.java
@@ -1,4 +1,4 @@
-package roomescape.repository;
+package roomescape.controller;
 
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
@@ -15,7 +15,7 @@ import static org.hamcrest.Matchers.is;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_EACH_TEST_METHOD)
-public class ApiTest {
+public class TimesApiTest {
 
     @LocalServerPort
     int port;
