+++ src/test/java/roomescape/repository/TimesCrudTest.java
@@ -14,7 +14,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_EACH_TEST_METHOD)
-public class CrudTest {
+public class TimesCrudTest {
 
     @Autowired
     private ReservationTimeRepository reservationTimeRepository;
